move node managers into dedicated dirs
