+++ src/lib/nodes/managers/EndpointManager.ts
@@ -1,4 +1,4 @@
-import ServiceMessage from '../messages/payloads/ServiceMessage';
+import ServiceMessage from '../../messages/payloads/ServiceMessage';
 import { BaseNodeManager } from './BaseNodeManager';
 
 export class EndpointManager extends BaseNodeManager<ServiceMessage> {}
